+++ src/lib/openai.js
