+++ src/platforms/abstract/abstractBuilder.js
@@ -1,0 +1,7 @@
+class AbstractBuilder {
+
+}
+
+module.exports = {
+    AbstractBuilder
+}
