+++ src/platforms/abstract/abstractParser.js
@@ -1,0 +1,7 @@
+class AbstractParser {
+
+}
+
+module.exports = {
+    AbstractParser
+}
