+++ src/platforms/github/githubBuilder.js
@@ -1,0 +1,9 @@
+const { AbstractBuilder } = require('../abstract/abstractBuilder');
+
+class GithubBuilder extends AbstractBuilder {
+
+}
+
+module.exports = {
+    GithubBuilder
+}
